src/main/java/Autor.java
import java.time.Instant;

public class Autor {

    private final String nome;
    private final String sobrenome;
    private final Instant data_nascimento;
    private final int id_autor;

    public Autor(String nome, String sobrenome, Instant data_nascimento, int id_autor) {
        this.nome = nome;
        this.sobrenome = sobrenome;
        this.data_nascimento = data_nascimento;
        this.id_autor = id_autor;
    }

    public String getNome() {
        return nome;
    }

    public String getSobrenome() {
        return sobrenome;
    }

    public Instant getData_nascimento() {
        return data_nascimento;
    }

    public int getId_autor() {
        return id_autor;
    }
}
